Removed commands and client_handler header from client and server files and icludes and src file added
